finished s3 and changed file system
