+++ .gitignore
@@ -5,3 +5,4 @@ secrets.js
 .DS_Store
 npm-debug.log
 yarn-error.log
+.env
