+++ package.json
@@ -47,6 +47,7 @@
     "axios": "^0.19.2",
     "compression": "^1.7.3",
     "connect-session-sequelize": "^6.1.1",
+    "dotenv": "^8.2.0",
     "express": "^4.16.4",
     "express-fileupload": "^1.1.7-alpha.3",
     "express-generator": "^4.16.1",
